midified precision double to float and added -1e-8 instead of 0 to add tolerance

+++ incs/camera.h
@@ -6,7 +6,7 @@
 /*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 15:45:00 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/12/12 14:01:36 by hmontoya         ###   ########.fr       */
+/*   Updated: 2024/12/17 11:39:45 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,13 +19,13 @@ typedef struct s_scene t_scene;
 
 typedef struct	s_camera
 {
-	double	fovH;
-	double	fovV;
+	float	fovH;
+	float	fovV;
 	t_vec3	axis;
 	t_p3	pos;
 	t_p3	center;
-	double	focal_length;
-	double	pixel_sample_scale;
+	float	focal_length;
+	float	pixel_sample_scale;
 	int		samples_per_pixel;
 }	t_camera;
