+++ incs/lights.h
@@ -6,7 +6,7 @@
 /*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/14 16:26:33 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/11/28 17:01:31 by hmontoya         ###   ########.fr       */
+/*   Updated: 2024/12/17 11:39:45 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,13 +22,13 @@
 
 typedef struct s_ambient
 {
-	double	range;
+	float	range;
 	t_color	rgb;
 }	t_ambient;
 
 typedef struct	s_light
 {
-	double	brghtnss;
+	float	brghtnss;
 	t_vec3	pos;
 	t_color	rgb;
 }	t_light;
