+++ srcs/math-convertions.c
@@ -6,7 +6,7 @@
 /*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/19 19:05:30 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/12/11 20:06:26 by hmontoya         ###   ########.fr       */
+/*   Updated: 2024/12/17 11:39:45 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 #include "maths.h"
 #include "vectors.h"
 
-double degrees_to_radians(double degrees)
+float degrees_to_radians(float degrees)
 {
     return degrees * PI / 180.0;
 }
