+++ srcs/math-interval-absolutes.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   math-interval-absolutes.c                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hmontoya <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 14:47:34 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/12/16 14:47:36 by hmontoya         ###   ########.fr       */
+/*   Updated: 2024/12/17 11:39:45 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ void empty(t_interval *inter)
 	inter->max = -INFINITY;	
 }
 
-double clamp(t_interval *inter, double x) 
+float clamp(t_interval *inter, float x) 
 {
 	if (x < inter->min)
 		return inter->min;
