+++ srcs/vec3-float-operations.c
@@ -1,7 +1,7 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   vec3-double-operations.c                           :+:      :+:    :+:   */
+/*   vec3-float-operations.c                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hmontoya <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
@@ -12,7 +12,7 @@
 
 #include "vectors.h"
 
-t_vec3		mult_v3_dbl(t_vec3 v1, double t)
+t_vec3		mult_v3_dbl(t_vec3 v1, float t)
 {
 	t_vec3 new;
 	new.x = v1.x * t;
@@ -21,7 +21,7 @@ t_vec3		mult_v3_dbl(t_vec3 v1, double t)
 	return (new);
 }
 
-t_vec3		div_v3_dbl(t_vec3 v1, double t)
+t_vec3		div_v3_dbl(t_vec3 v1, float t)
 {
 	t_vec3 new;
 	new.x = (1/t) * v1.x;
@@ -30,7 +30,7 @@ t_vec3		div_v3_dbl(t_vec3 v1, double t)
 	return (new);
 }
 
-t_vec3		sum_v3_dbl(t_vec3 v1, double t)
+t_vec3		sum_v3_dbl(t_vec3 v1, float t)
 {
 	t_vec3 new;
 	new.x = v1.x + t;
@@ -39,7 +39,7 @@ t_vec3		sum_v3_dbl(t_vec3 v1, double t)
 	return (new);
 }
 
-t_vec3		rest_v3_dbl(t_vec3 v1, double t)
+t_vec3		rest_v3_dbl(t_vec3 v1, float t)
 {
 	t_vec3 new;
 	new.x = v1.x - t;
@@ -48,8 +48,8 @@ t_vec3		rest_v3_dbl(t_vec3 v1, double t)
 	return (new);
 }
 
-double		vec3_to_dbl(t_vec3 v)
+float		vec3_to_dbl(t_vec3 v)
 {
 	return (v.x + v.y + v.z);
 }
-//return ((double)(v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z));
+//return ((float)(v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z));
